Chapter 5.6. Custom template functions
Pipelining
In the code above, we called our custom template function like this:

<time>Created: {{humanDate .Created}}</time>
An alternative approach is to use the | character to pipeline values to a function. This works a bit like pipelining outputs from one command to another in Unix terminals. We could re-write the above as:

<time>Created: {{.Created | humanDate}}</time>
A nice feature of pipelining is that you can make an arbitrarily long chain of template functions which use the output from one as the input for the next. For example, we could pipeline the output from our humanDate function to the inbuilt printf function like so:

<time>{{.Created | humanDate | printf "Created: %s"}}</time>